Backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
firebase-admin
bson